webpack babel

+++ src/actions/auth.js
@@ -9,7 +9,7 @@ import {
 	SEND_RESET_EMAIL_FAIL,
 } from './types';
 
-import { serviceGet, servicePost, setAuthToken } from '../helpers';
+import { serviceGet, servicePost, setAuthToken, isClient } from '../helpers';
 import { errorToast, successToast } from './toast';
 
 // Load User : Every time we logged in or register or refresh the page its gonna load.
@@ -18,17 +18,17 @@ export const loadUser = () => async (dispatch) => {
 	dispatch({
 		type: 'SET_AUTH_LOADER',
 	});
-	if (localStorage.token) {
+	if (isClient && localStorage.token) {
 		setAuthToken(localStorage.token);
 	}
 
-	if (localStorage.userId) {
+	if (isClient && localStorage.userId) {
 		const userId = localStorage.getItem('userId');
 
 		try {
 			const res = await serviceGet(`api/auth/user/${userId}`);
 			// displayToast
-			successToast(res)
+			successToast(res);
 
 			dispatch({
 				type: USER_LOADED,
@@ -36,7 +36,7 @@ export const loadUser = () => async (dispatch) => {
 			});
 		} catch (err) {
 			// display toast
-			errorToast(err)
+			errorToast(err);
 			dispatch({
 				type: AUTH_ERROR,
 			});
@@ -48,11 +48,11 @@ export const refreshUser = (user) => async (dispatch) => {
 	dispatch({
 		type: 'SET_AUTH_LOADER',
 	});
-	if (localStorage.token) {
+	if (isClient && localStorage.token) {
 		setAuthToken(localStorage.token);
 	}
 
-	if (localStorage.userId) {
+	if (isClient && localStorage.userId) {
 		const userId = localStorage.getItem('userId');
 
 		try {
@@ -89,13 +89,13 @@ export const register = (data) => async (dispatch) => {
 			payload: res.data,
 		});
 
-		successToast(res)
+		successToast(res);
 
 		console.log(res);
 		dispatch(loadUser());
 	} catch (err) {
 		// const errors = err.response.data.errors;
-		errorToast(err)
+		errorToast(err);
 		dispatch({
 			type: REGISTER_FAIL,
 		});
@@ -104,39 +104,39 @@ export const register = (data) => async (dispatch) => {
 
 export const login =
 	({ email, password }) =>
-		async (dispatch) => {
-			dispatch({
-				type: 'SET_AUTH_LOADER',
-			});
-			const headers = {
-				'Content-Type': 'application/json',
-			};
-
-			const body = JSON.stringify({ email, password });
-
-			try {
-				const res = await servicePost('api/auth/login', body, headers);
-
-				console.log(res);
-
-				dispatch({
-					type: res.status === 1 ? LOGIN_SUCCESS : LOGIN_FAIL,
-					payload: { userId: res.data?.user?.userId, token: res.data?.token },
-				});
-
-				dispatch(loadUser());
-				successToast(res)
-			} catch (err) {
-				console.log(err);
-				errorToast(err)
-				dispatch({
-					type: LOGIN_FAIL,
-				});
-			}
+	async (dispatch) => {
+		dispatch({
+			type: 'SET_AUTH_LOADER',
+		});
+		const headers = {
+			'Content-Type': 'application/json',
 		};
 
+		const body = JSON.stringify({ email, password });
+
+		try {
+			const res = await servicePost('api/auth/login', body, headers);
+
+			console.log(res);
+
+			dispatch({
+				type: res.status === 1 ? LOGIN_SUCCESS : LOGIN_FAIL,
+				payload: { userId: res.data?.user?.userId, token: res.data?.token },
+			});
+
+			dispatch(loadUser());
+			successToast(res);
+		} catch (err) {
+			console.log(err);
+			errorToast(err);
+			dispatch({
+				type: LOGIN_FAIL,
+			});
+		}
+	};
+
 export const logout = () => (dispatch) => {
-	localStorage.clear();
+	if (isClient) localStorage.clear();
 	dispatch({
 		type: 'LOGOUT',
 	});
@@ -144,30 +144,30 @@ export const logout = () => (dispatch) => {
 
 export const sendResetEmail =
 	({ email }) =>
-		async (dispatch) => {
-			const config = {
-				headers: {
-					'Content-Type': 'application/json',
-				},
-			};
-			const body = JSON.stringify({ email });
-
-			try {
-				const res = await servicePost('/api/forgotpassword', body, config);
-
-				const messagesArray = res.data.messages;
-				// brand added message alert
-				messagesArray.forEach((message) => dispatch((message.msg, 'danger')));
-				dispatch({
-					type: SEND_RESET_EMAIL,
-					payload: res.data,
-				});
-			} catch (err) {
-				const errors = err && err.response.data.errors;
-				if (errors) {
-				}
-				dispatch({
-					type: SEND_RESET_EMAIL_FAIL,
-				});
-			}
+	async (dispatch) => {
+		const config = {
+			headers: {
+				'Content-Type': 'application/json',
+			},
 		};
+		const body = JSON.stringify({ email });
+
+		try {
+			const res = await servicePost('/api/forgotpassword', body, config);
+
+			const messagesArray = res.data.messages;
+			// brand added message alert
+			messagesArray.forEach((message) => dispatch((message.msg, 'danger')));
+			dispatch({
+				type: SEND_RESET_EMAIL,
+				payload: res.data,
+			});
+		} catch (err) {
+			const errors = err && err.response.data.errors;
+			if (errors) {
+			}
+			dispatch({
+				type: SEND_RESET_EMAIL_FAIL,
+			});
+		}
+	};
